requirements.txt
PyNaCl
discord.py
google-api-python-client
python-dotenv
pytube
soundcloud-lib
spotipy
spotdl
streamlit
